Fix for Azure certificates compatibility
When running a CentOS container on a Debian based virtual machine on Azure (default), an error was thrown while mounting the certificates. This was because we tried to add them by default. This PR changes the behaviour so this no longer happens. A user can still renable these options using the `container.runOptions` as per any other Docker runner.

Signed-off-by: Adam Talbot <12817534+adamrtalbot@users.noreply.github.com>